NLPformed/requirements.txt
easyocr==1.7.1
torch
torchvision
torchaudio
Pillow
opencv-python-headless
